requirements.txt
python-nmap==0.6.1
pyshark==0.5.3
scapy==2.4.5
wifi==0.3.3
requests==2.31.0
PyYAML==6.0
netifaces==0.11.0
tabulate==0.9.0